IOT-119: IOT Cache Requirements   - get and is in ExpiryPolicy.java   - AbstractCache.java with ExpiryPolicy.java   - Make classes extend AbstractCache.java   - Added getExpiryPolicy to Cache interface

+++ core/src/main/java/com/hortonworks/iotas/cache/view/redis/DataStoreBackedCache.java
@@ -18,12 +18,15 @@
 
 package com.hortonworks.iotas.cache.view.redis;
 
+import com.hortonworks.iotas.cache.AbstractCache;
 import com.hortonworks.iotas.cache.Cache;
 import com.hortonworks.iotas.cache.stats.CacheStats;
 import com.hortonworks.iotas.cache.view.datastore.DataStore;
 import com.hortonworks.iotas.cache.view.datastore.writer.DataStoreWriter;
 import com.hortonworks.iotas.cache.view.loader.CacheLoader;
 import com.hortonworks.iotas.storage.exception.StorageException;
+import com.sun.javaws.exceptions.InvalidArgumentException;
+import com.sun.org.apache.bcel.internal.generic.DASTORE;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -34,7 +37,7 @@ import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.Future;
 
-public class DataStoreBackedCache<K,V> implements Cache<K,V> {
+public class DataStoreBackedCache<K,V> extends AbstractCache<K,V> implements Cache<K,V> {
     private static final Logger LOG = LoggerFactory.getLogger(DataStoreBackedCache.class);
     private final Cache<K, V> cache;
     private final CacheLoader<K, V> cacheLoader;
@@ -43,6 +46,9 @@ public class DataStoreBackedCache<K,V> implements Cache<K,V> {
 
     public DataStoreBackedCache(Cache<K,V> cache, DataStore<K, V> dataStore,
             CacheLoader<K,V> cacheLoader, DataStoreWriter<K, V> dataStoreWriter) {
+
+        validateArguments(cache, dataStore, cacheLoader, dataStoreWriter);
+
         this.cache = cache;
         this.dataStore = dataStore;
         this.cacheLoader = cacheLoader;
@@ -65,7 +71,7 @@ public class DataStoreBackedCache<K,V> implements Cache<K,V> {
     @Override
     public V get(K key) throws StorageException {
         V val = cache.get(key);
-        if (val == null) {
+        if (val == null && dataStore != null) {     // in sync read through
             val = dataStore.read(key);
         }
         return val;
@@ -74,19 +80,23 @@ public class DataStoreBackedCache<K,V> implements Cache<K,V> {
     @Override
     public void put(K key, V val) {
         cache.put(key, val);
-        dataStoreWriter.write(key, val);
+        if (dataStoreWriter != null) {              // in sync write through
+            dataStoreWriter.write(key, val);
+        }
     }
 
     @Override
     public Map<K, V> getAll(Collection<? extends K> keys) {  // TODO what if trying to load more keys than max number of keys that be kept in the cache ?
         Map<K, V> present = cache.getAll(keys);
-        if (present == null || present.isEmpty()) {
-            present = dataStore.readAll(keys);                  // in sync read through
-        } else if (present.size() < keys.size()) {
-            Set<K> notPresent = new HashSet<>(keys);
-            notPresent.removeAll(present.keySet());
-            Map<K, V> loaded = dataStore.readAll(notPresent);   // in sync read through
-            present.putAll(loaded);
+        if (dataStore != null) {
+            if (present == null || present.isEmpty()) {
+                present = dataStore.readAll(keys);                  // in sync read through
+            } else if (present.size() < keys.size()) {
+                Set<K> notPresent = new HashSet<>(keys);
+                notPresent.removeAll(present.keySet());
+                Map<K, V> loaded = dataStore.readAll(notPresent);   // in sync read through
+                present.putAll(loaded);
+            }
         }
 //        LOG.debug("Entries existing in cache [{}]. Keys non existing in cache: [{}]", present, notPresent.removeAll(loaded));
         return present;
@@ -95,25 +105,31 @@ public class DataStoreBackedCache<K,V> implements Cache<K,V> {
     @Override
     public void putAll(Map<? extends K, ? extends V> entries) {
         cache.putAll(entries);
-        dataStoreWriter.writeAll(entries);
+        if (dataStoreWriter != null) {      // sync or async write, depending on the writing strategy chosen
+            dataStoreWriter.writeAll(entries);
+        }
     }
 
     @Override
     public void remove(K key) {
         cache.remove(key);
-        dataStoreWriter.delete(key);
+        if (dataStoreWriter != null) {      // sync or async delete, depending on the writing strategy chosen
+            dataStoreWriter.delete(key);
+        }
     }
 
     @Override
     public void removeAll(Collection<? extends K> keys) {
         cache.removeAll(keys);
-        dataStoreWriter.deleteAll(keys);
+        if (dataStoreWriter != null) {      // sync or async delete, depending on the writing strategy chosen
+            dataStoreWriter.deleteAll(keys);
+        }
     }
 
     @Override
     public void clear() {
         cache.clear();
-        LOG.info("Entries only removed from cache but not from backing data store");    //TODO Remove from DB as well ?
+        LOG.info("Entries only removed from cache but not from backing data store");    //TODO: Do we want to remove from DB as well ?
     }
 
     @Override
@@ -123,6 +139,39 @@ public class DataStoreBackedCache<K,V> implements Cache<K,V> {
 
     @Override
     public CacheStats stats() {
-        return null;
+        return cache.stats();
     }
+
+    @Override
+    public String toString() {
+        return "DataStoreBackedCache{" +
+                "cache=" + cache +
+                ", cacheLoader=" + cacheLoader +
+                ", dataStoreWriter=" + dataStoreWriter +
+                ", dataStore=" + dataStore +
+                "} " + super.toString();
+    }
+
+    // =========== Private helper methods ===========
+
+    private void validateArguments(Cache<K, V> cache, DataStore<K, V> dataStore,
+                                   CacheLoader<K, V> cacheLoader, DataStoreWriter<K, V> dataStoreWriter) {
+        if (cache == null) {
+            throw new IllegalArgumentException("Cache reference cannot be null");
+        }
+
+        if (dataStore == null && cacheLoader == null && dataStoreWriter == null) {
+            throw new IllegalArgumentException(String.format("At least one non null implementation of %s,  %s, or  %s " +
+                            "is required. If no backing data store is required consider using a non backed implementation of %s",
+                    getSimpleName(DataStore.class), getSimpleName(DataStoreWriter.class),
+                    getSimpleName(CacheLoader.class), getSimpleName(Cache.class)));
+        }
+
+        LOG.info("Created {}", this);
+    }
+
+    private String getSimpleName(Class<?> clazz) {
+        return clazz.getSimpleName();
+    }
+
 }
